requirements.txt
gunicorn
flask
python-dotenv
flask-jwt-extended
flask-cors
pymongo
flask-restful